tweak outputs from base, ssmcleanup
