+++ v2/ssm_agent/cdk/bin/app.ts
@@ -3,16 +3,16 @@
 
 import "source-map-support/register"
 import * as cdk from "@aws-cdk/core"
-import { BaseImplementationStack } from "../lib/BaseImplementationStack"
+import { SsmComponentStack } from "../lib/SsmComponentStack"
 
 const app = new cdk.App()
-const name = process.env.STACK_NAME || "gg-accel-base"
+const name = process.env.STACK_NAME || "gg-accel-ssm"
 
-new BaseImplementationStack(app, name, {
+new SsmComponentStack(app, name, {
   env: {
     account: process.env.CDK_DEPLOY_ACCOUNT || process.env.CDK_DEFAULT_ACCOUNT,
     region: process.env.CDK_DEPLOY_REGION || process.env.CDK_DEFAULT_REGION
   },
   stackName: name,
-  description: "Greengrass Accelerator v2 - Base Implementation for development"
+  description: "Greengrass Accelerator v2 - SSM component"
 })
