--- v2/ssm_agent/cdk/lib/GreengrassV2Component/assets/Pipfile
@@ -1,12 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-pyyaml = "==5.4.1"
-
-[dev-packages]
-
-[requires]
-python_version = "3.8"
