--- v2/ssm_agent/cdk/lib/IotThingCertPolicy/assets/Pipfile
@@ -1,12 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-cryptography = "==3.4.7"
-
-[dev-packages]
-
-[requires]
-python_version = "3.8"
