--- v2/ssm_agent/cdk/lib/IotThingCertPolicy/assets/requirements.txt
@@ -1,1 +1,0 @@
-cryptography==3.4.7
