+++ v2/ssm_agent/cdk/lib/SsmComponentStack.ts
@@ -1,0 +1,63 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: MIT-0
+
+import * as path from "path"
+import * as cdk from "@aws-cdk/core"
+
+export class SsmComponentStack extends cdk.Stack {
+  constructor(scope: cdk.Construct, id: string, props?: cdk.StackProps) {
+    super(scope, id, props)
+
+    if (cdk.Stack.of(this).stackName.length > 20) {
+      console.error("Stack name must be less than 20 characters in length")
+      process.exitCode = 1
+    }
+    const parentStack = this.node.tryGetContext("baseStack")
+
+    // suffix to use for all stack resources to make unique
+    // In this stack all resources will use the format STACKNAME-RESOURCENAME-RANDOMSUFFIX
+    const stackRandom: string = makeid(8)
+
+    // Load parameters from parent stack
+    const thingArn = cdk.Fn.importValue(`${parentStack}-ThingArn`)
+    const certificateArn = cdk.Fn.importValue(`${parentStack}-CertificateArn`)
+    const iamRoleArn = cdk.Fn.importValue(`${parentStack}-IamRoleArn`)
+
+    // Layered constructs - each constructs derived values can be used for subsequent constructs
+
+    // update iot policy with any needed permissions -- remove policy permissions or just leave?
+    // add inline policy to iamRole -- remove inline policy by name
+    // create thingGroup and add thing to it -- delete group
+    // create components -- delete component
+    // create deployment -- cancel deployment
+
+    // ************ End of CDK Constructs / stack - Supporting functions below ************
+
+    function makeid(length: number) {
+      // Generate a n-length random value for each resource
+      var result = ""
+      var characters = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789"
+      var charactersLength = characters.length
+      for (var i = 0; i < length; i++) {
+        result += characters.charAt(Math.floor(Math.random() * charactersLength))
+      }
+      return result
+    }
+
+    interface ResourceName {
+      stackName: string
+      baseName: string
+      suffix: string
+      resourceRegex: string
+      maxLength: number
+    }
+
+    function fullResourceName({ stackName, baseName, suffix, resourceRegex, maxLength }: ResourceName) {
+      let re = new RegExp(`[^\\[${resourceRegex}]`, "g")
+      let resourceName = `${stackName}-${baseName}`.replace(re, "")
+      resourceName = resourceName.substring(0, maxLength - suffix.length - 1)
+      resourceName = `${resourceName}-${suffix}`
+      return resourceName
+    }
+  }
+}
