+++ v2/ssm_agent/cdk/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "ggv2-accelerator-base",
-  "version": "0.5.0",
+  "name": "ggv2-accelerator-ssm-component",
+  "version": "0.1.0",
   "license": "MIT",
   "bin": {
     "cdk": "bin/app.js"
@@ -24,20 +24,8 @@
     "typescript": "~4.3.5"
   },
   "dependencies": {
-    "@aws-cdk/aws-cloudwatch": "1.115.0",
-    "@aws-cdk/aws-iam": "1.115.0",
-    "@aws-cdk/aws-iot": "1.115.0",
-    "@aws-cdk/aws-lambda": "1.115.0",
-    "@aws-cdk/aws-lambda-python": "1.115.0",
-    "@aws-cdk/aws-logs": "1.115.0",
-    "@aws-cdk/aws-s3": "1.115.0",
-    "@aws-cdk/aws-s3-assets": "1.115.0",
     "@aws-cdk/core": "1.115.0",
-    "@aws-cdk/custom-resources": "1.115.0",
-    "archiver": "^5.3.0",
-    "lodash": "^4.17.21",
-    "source-map-support": "^0.5.16",
-    "tempy": "^1.0.1"
+    "source-map-support": "^0.5.16"
   },
   "prettier": {
     "tabWidth": 2,
